requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
bidict==0.22.0
cachelib==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
DateTime==4.4
dnspython==2.2.1
eventlet==0.33.1
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Session==0.4.0
Flask-SocketIO==5.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pygame==2.1.2
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==1.5
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.1
zope.interface==5.4.0
